+++ sig-provider/src/lib.rs
@@ -1,12 +1,10 @@
 mod aggregator;
 mod proto;
-mod provider;
 mod server;
 mod settings;
 mod sources;
 
 pub use aggregator::SignatureAggregator;
-pub use provider::SignatureProvider;
 pub use server::sig_provider;
 pub use settings::Settings;
 pub use sources::*;
